Untagle BinPackingNodeAllocatorService and ExponentialGrowthPartitionMemoryEstimator
